Fixed QM scripts for Abyssea-Uleguerand
