+++ scripts/zones/Abyssea-Uleguerand/npcs/qm4.lua
@@ -12,7 +12,7 @@ require("scripts/globals/status");
 
 function onTrade(player,npc,trade)
 --[[
-    if (trade:hasItemQty(3259,1) and trade:getItemCount() == 1) then -- Player has all the required items.
+    if (trade:hasItemQty(3259,1) and trade:hasItemQty(3249,1) and trade:getItemCount() == 2) then -- Player has all the required items.
         if (GetMobAction(17813931) == ACTION_NONE) then -- Mob not already spawned from this
             SpawnMob(17813931):updateClaim(player); -- Spawn NM, Despawn after inactive for 5 minutes (pt has to reclaim within 5 of a wipe)
             player:tradeComplete();
@@ -26,7 +26,7 @@ end;
 -----------------------------------
 
 function onTrigger(player,npc)
-    player:startEvent(1010, 3259); -- Inform player what items they need.
+    player:startEvent(1010, 3259, 3249); -- Inform player what items they need.
 end;
 
 -----------------------------------
